requirements.txt
torch == 1.10.0
gpytorch == 1.5.1
numpy >= 1.16.4
scipy >= 1.4.1
scikit-learn == 0.24.2
pyro-ppl == 1.7.0